Add files via upload
The Lorentz system can be controlled using linear quadratic Gaussian (LQG) control, which is a type of linear control method. The LQG controller can be designed using the linear quadratic regulator (LQR) technique, which involves linearizing the system about a desired operating point and using linear control techniques to regulate the system.

In this simulation, we will design an LQG controller for a Lorentz system and simulate its behavior in Python. We will use the Scipy and Numpy libraries for the simulation.
In this code, we first define the Lorentz system using the state space representation, where x is the state vector and u is the control input. Then, we design an LQR controller using the solve_continuous_are function from the Scipy library and compute the feedback gain K.

Next, we design an LQG controller using the LQR design and the Kalman filter. We use the lqr function from the Scipy library to compute the Kalman gain L.

Finally, we simulate the closed-loop system using the odeint function from the Scipy library, which integrates the system over